Cambios el 24/03/2017

+++ UIF/Web/DTOs/Venta/ConsultaOrdenTrabajoVistaDTOs.cs
@@ -49,7 +49,7 @@ namespace Web.DTOs.Venta
         /// EstadoPago
         /// </summary>
         
-        public int? Cantidad { get; set; }
+        public decimal? Cantidad { get; set; }
 
         /// <summary>
         /// Prenda
